can now rewrite to uint8 with BitSet256

+++ table_test.go
@@ -2591,7 +2591,7 @@ func TestLastIdxLastBits(t *testing.T) {
 	tests := []struct {
 		pfx       netip.Prefix
 		wantDepth int
-		wantBits  int
+		wantBits  uint8
 	}{
 		{
 			pfx:       mpp("0.0.0.0/0"),
